tests/smoke/requirements.txt
# Smoke Test Dependencies for Sources Sought AI

# Core testing framework
requests>=2.28.0
pytest>=7.0.0
pytest-asyncio>=0.21.0

# AWS integration
boto3>=1.26.0
botocore>=1.29.0

# Docker integration
docker>=6.0.0

# Redis integration
redis>=4.5.0

# JWT handling
PyJWT>=2.6.0

# Data handling
aiofiles>=22.0.0

# Monitoring and metrics
psutil>=5.9.0

# Development dependencies
black>=22.0.0
isort>=5.12.0
mypy>=1.0.0